arch-review: extract routes_mcp.py from mcp_server.py
Move the six /api/mcp/* endpoint functions (status, usage, env keys,
servers collection, server item, tools) and the two live tool-discovery
probe helpers (_probe_http_mcp_tools, _observed_project_mcp_tools) out of
ciao/mcp_server.py into ciao/web/routes_mcp.py, so HTTP handlers all live
under ciao/web/routes_*.py and mcp_server.py is the MCP adapter again. The
service methods still probe through the helpers, so mcp_server.py imports
them from routes_mcp (routes_mcp needs nothing from the adapter, so there
is no cycle; endpoints reach the service via request.app.state.mcp_service).
app.py imports the mcp endpoints from routes_mcp.
